rasp-rain-preview/requirements.txt
numpy
paho-mqtt
tensorflow